Add blog functionality with react-router-dom
The core functionality was extended to include a blog component. This update includes installing react-router-dom to handle routing. Each project now links to a separate blog post. BlogNavbar component was also added to navigate between different blog posts. In addition, several adjustments were made to existing components to improve the structure and clarity of the code.